+++ modules/handle/builtin/cpp/cancelBuiltin.cpp
@@ -1,0 +1,24 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "cancelBuiltin.hpp"
+#include "InputOutputArgumentsCheckers.hpp"
+#include "OverloadRequired.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::HandleGateway::cancelBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    nargincheck(argIn, 1, 1);
+    nargoutcheck(nLhs, 0, 1);
+    OverloadRequired("cancel");
+    return {};
+}
+//=============================================================================