+++ modules/handle/builtin/include/cancelBuiltin.hpp
@@ -9,19 +9,12 @@
 //=============================================================================
 #pragma once
 //=============================================================================
-#include "Evaluator.hpp"
-#include "nlsFunctions_manager_exports.h"
+#include "ArrayOf.hpp"
 //=============================================================================
-namespace Nelson {
+namespace Nelson::HandleGateway {
 //=============================================================================
-NLSFUNCTIONS_MANAGER_IMPEXP bool
-ClearBuiltin(const std::wstring& builtinName);
-//=============================================================================
-NLSFUNCTIONS_MANAGER_IMPEXP bool
-ClearAllBuiltin();
-//=============================================================================
-NLSFUNCTIONS_MANAGER_IMPEXP bool
-ClearMacroCache(Evaluator* eval);
+ArrayOfVector
+cancelBuiltin(int nLhs, const ArrayOfVector& argIn);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================