+++ modules/interpreter/builtin/cpp/onCleanup_usedBuiltin.cpp
@@ -1,0 +1,27 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "onCleanup_usedBuiltin.hpp"
+#include "OnCleanupObjectHandle.hpp"
+#include "Error.hpp"
+#include "usedHandle.hpp"
+#include "InputOutputArgumentsCheckers.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::InterpreterGateway::onCleanup_usedBuiltin(int nLhs, const ArrayOfVector& argIn)
+{
+    nargincheck(argIn, 0, 0);
+    nargoutcheck(nLhs, 0, 1);
+    ArrayOfVector retval(1);
+    retval << usedHandle(NLS_HANDLE_ONCLEANUP_CATEGORY_STR);
+    return retval;
+}
+//=============================================================================